Fix macOS build: cast libc::S_IFLNK to u32
On macOS, libc defines S_IFLNK as u16 while Linux uses u32.
Add explicit casts at all usage sites.

+++ crates/ferrosync-core/src/engine/transfer.rs
@@ -826,7 +826,7 @@ fn collect_directory(
 /// Get the platform's S_IFLNK value.
 #[cfg(unix)]
 fn libc_s_iflnk() -> u32 {
-    libc::S_IFLNK
+    libc::S_IFLNK as u32
 }
 
 #[cfg(not(unix))]
